pyproject.toml
[tool.poetry]
name = "funding"
version = "0.2.0"
description = ""
authors = ["VasyaPup"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<4"
httpx = ">=0.27"
python-dotenv = ">=1.0"
tzdata = ">=2024.1"
cryptography = "44.0.2"
eth-account = "0.13.7"
msgpack = "*"
eth-hash = { extras = ["pycryptodome"], version = "*" }
apexpro = "*"
lighter-sdk = "*"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0"
pytest-asyncio = ">=0.23"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
